Compress the final CSS files + remove CSS comments
The final CSS files are compiled in their full version. To trim down the final files and optimize resources, the native node-sass outputStyle option is added with a 'compressed' value, which removes all spaces and comments.

The savings is about ~5% over the non-compressed CSS files.

Also, in the assets/scss/_bootstrap.scss file, removed native CSS commenting syntax in favor of Sass commenting syntax so that it doesn't get compiled in our final CSS file.